move webrtc constants in namespace
