+++ packages/common/src/BrowserSession.ts
@@ -4,7 +4,7 @@ import BaseCall from './webrtc/BaseCall'
 import { ICacheDevices, IAudioSettings, IVideoSettings, BroadcastParams, SubscribeParams } from './util/interfaces'
 import { registerOnce, trigger } from './services/Handler'
 import { SwEvent, SESSION_ID, DeviceType } from './util/constants'
-import { State } from './util/constants/call'
+import { State } from './webrtc/constants'
 import { getDevices, scanResolutions, removeUnsupportedConstraints, checkDeviceIdConstraints, destructSubscribeResponse, getUserMedia } from './webrtc/helpers'
 import { findElementByType } from './util/helpers'
 import { Unsubscribe, Subscribe, Broadcast } from './messages/Verto'
