+++ packages/common/src/webrtc/constants.ts
@@ -1,4 +1,4 @@
-import { CallOptions } from '../interfaces'
+import { CallOptions } from '../util/interfaces'
 
 export const DEFAULT_CALL_OPTIONS: CallOptions = {
   destinationNumber: '',
@@ -51,4 +51,42 @@ export enum ConferenceAction {
   LayoutInfo = 'layoutInfo',
   LayoutList = 'layoutList',
   ModCmdResponse = 'modCommandResponse',
+}
+
+export enum PeerType {
+  Offer = 'offer',
+  Answer = 'answer'
+}
+
+export enum Direction {
+  Inbound = 'inbound',
+  Outbound = 'outbound'
+}
+
+export enum VertoMethod {
+  Invite = 'verto.invite',
+  Attach = 'verto.attach',
+  Answer = 'verto.answer',
+  Info = 'verto.info',
+  Display = 'verto.display',
+  Media = 'verto.media',
+  Event = 'verto.event',
+  Bye = 'verto.bye',
+  Punt = 'verto.punt',
+  Broadcast = 'verto.broadcast',
+  Subscribe = 'verto.subscribe',
+  Unsubscribe = 'verto.unsubscribe',
+  ClientReady = 'verto.clientReady',
+  Modify = 'verto.modify',
+}
+
+export const NOTIFICATION_TYPE = {
+  generic: 'event',
+  [VertoMethod.Display]: 'participantData',
+  [VertoMethod.Attach]: 'participantData',
+  conferenceUpdate: 'conferenceUpdate',
+  callUpdate: 'callUpdate',
+  vertoClientReady: 'vertoClientReady',
+  userMediaError: 'userMediaError',
+  refreshToken: 'refreshToken',
 }
